legacy_data/gid8_training_package/requirements.txt
torch==2.3
transformers
google-cloud-storage